Move kubeconfig funcs to kube_context
